Functionality to re-use a local repo
Fixes #4
