+++ bin/dogit
@@ -6,7 +6,7 @@ require file_exists(__DIR__ . '/../vendor/autoload.php')
     : __DIR__ . '/../../../../vendor/autoload.php';
 
 use dogit\Commands\CloneProjectMergeRequest;
-use dogit\Commands\IssueClone;
+use dogit\Commands\IssueMergeRequest;
 use dogit\Commands\ProjectCloneCommand;
 use dogit\Commands\GitCommand;
 use dogit\Commands\IssueTimelineCommand;
@@ -16,7 +16,7 @@ $application = new Application();
 
 $application->add(new CloneProjectMergeRequest());
 $application->add(new GitCommand());
-$application->add(new IssueClone());
+$application->add(new IssueMergeRequest());
 $application->add(new IssueTimelineCommand());
 $application->add(new ProjectCloneCommand());
