+++ src/Commands/IssueMergeRequest.php
@@ -1,0 +1,197 @@
+<?php
+
+declare(strict_types=1);
+
+namespace dogit\Commands;
+
+use CzProject\GitPhp\Git;
+use CzProject\GitPhp\GitException;
+use dogit\Commands\Options\IssueMergeRequestOptions;
+use dogit\DrupalOrg\DrupalApi;
+use dogit\DrupalOrg\DrupalOrgObjectRepository;
+use dogit\DrupalOrg\IssueGraph\DrupalOrgIssueGraph;
+use dogit\DrupalOrg\IssueGraph\Events\IssueEvent;
+use dogit\DrupalOrg\IssueGraph\Events\MergeRequestCreateEvent;
+use dogit\Git\CliRunner;
+use dogit\Git\GitOperator;
+use Symfony\Component\Console\Command\Command;
+use Symfony\Component\Console\Input\InputArgument;
+use Symfony\Component\Console\Input\InputInterface;
+use Symfony\Component\Console\Input\InputOption;
+use Symfony\Component\Console\Logger\ConsoleLogger;
+use Symfony\Component\Console\Output\OutputInterface;
+use Symfony\Component\Console\Style\SymfonyStyle;
+
+/**
+ * Clones or checks out an existing Merge request for an issue.
+ */
+final class IssueMergeRequest extends Command
+{
+    use Traits\HttpTrait;
+
+    protected static $defaultName = 'issue:clone';
+
+    protected function configure(): void
+    {
+        $this
+            ->setDescription('Interactively check out a MR for an issue.')
+            ->addArgument(IssueMergeRequestOptions::ARGUMENT_ISSUE_ID, InputArgument::REQUIRED)
+            ->addArgument(IssueMergeRequestOptions::ARGUMENT_DIRECTORY, InputArgument::REQUIRED)
+            ->addOption(IssueMergeRequestOptions::OPTION_COOKIE, null, InputOption::VALUE_OPTIONAL | InputOption::VALUE_IS_ARRAY, 'Add cookies to HTTP requests.', [])
+            ->addOption(IssueMergeRequestOptions::OPTION_HTTP, null, InputOption::VALUE_NONE, 'Use HTTP instead of SSH.')
+            ->addOption(IssueMergeRequestOptions::OPTION_SINGLE, 's', InputOption::VALUE_NONE, '(no interaction) If there is only one merge request then check it out without prompting. If multiple merge requests are found then the command will quit.')
+            ->addOption(IssueMergeRequestOptions::OPTION_NO_CACHE, null, InputOption::VALUE_NONE, 'Whether to not use a HTTP cache.');
+    }
+
+    protected function execute(InputInterface $input, OutputInterface $output): int
+    {
+        $io = new SymfonyStyle($input, $output);
+        $logger = new ConsoleLogger($io);
+        $options = IssueMergeRequestOptions::fromInput($input);
+        $repository = new DrupalOrgObjectRepository();
+        [$httpFactory, $httpAsyncClient] = $this->http($logger, $options->noHttpCache, $options->cookies);
+        $api = new DrupalApi($httpFactory, $httpAsyncClient, $repository);
+
+        $issue = $api->getIssue($options->nid);
+        $io->text(sprintf(
+            '<href=%s>Issue #%s for %s at %s: %s</>',
+            $issue->url(),
+            $issue->id(),
+            $issue->getProjectName(),
+            $issue->getCreated()->format('r'),
+            $issue->getTitle(),
+        ));
+
+        /** @var \dogit\DrupalOrg\IssueGraph\Events\IssueEventInterface[] $events */
+        $events = iterator_to_array((new DrupalOrgIssueGraph(
+            $httpFactory,
+            $httpAsyncClient,
+            $repository,
+            $issue->url(),
+        ))->graph(), false);
+
+        $mergeRequestCreateEvents = IssueEvent::filterMergeRequestCreateEvents($events);
+        if (0 === count($mergeRequestCreateEvents)) {
+            $io->error('No merge requests found.');
+
+            return Command::FAILURE;
+        }
+
+        // Remap to MR IDs.
+        /** @var \dogit\DrupalOrg\IssueGraph\Events\MergeRequestCreateEvent[] $mergeRequestCreateEvents */
+        $mergeRequestCreateEvents = array_combine(
+            array_map(
+                fn (MergeRequestCreateEvent $mergeRequestCreateEvent): int => $mergeRequestCreateEvent->mergeRequestId(),
+                $mergeRequestCreateEvents,
+            ),
+            $mergeRequestCreateEvents,
+        );
+
+        $choices = array_map(
+            fn (MergeRequestCreateEvent $mergeRequestCreateEvent): string => sprintf('Merge request !%d: %s from comment #%d', $mergeRequestCreateEvent->mergeRequestId(), $mergeRequestCreateEvent->getGitBranch(), $mergeRequestCreateEvent->getComment()->getSequence()),
+            $mergeRequestCreateEvents,
+        );
+
+        if ($options->single) {
+            if (1 === count($mergeRequestCreateEvents)) {
+                $io->note('Found a single MR.');
+                $mergeRequestCreateEvent = reset($mergeRequestCreateEvents);
+            } else {
+                $io->error('Multiple merge requests found. Re-run this command without --single.');
+
+                return Command::FAILURE;
+            }
+        } else {
+            $response = $io->choice(
+                'Please select merge request to checkout',
+                $choices,
+                1 === count($choices) ? array_key_first($choices) : 0,
+            );
+            $mrId = array_search($response, $choices, true);
+            $mergeRequestCreateEvent = $mergeRequestCreateEvents[$mrId];
+        }
+
+        $io->note(sprintf(
+            'Checking out merge request !%d: %s from comment #%d into directory %s',
+            $mergeRequestCreateEvent->mergeRequestId(),
+            $mergeRequestCreateEvent->getGitBranch(),
+            $mergeRequestCreateEvent->getComment()->getSequence(),
+            $options->directory,
+        ));
+
+        $runner = new CliRunner();
+        $git = new Git($runner);
+
+        // If this is an existing repo.
+        try {
+            $gitIo = GitOperator::fromDirectory($options->directory);
+            $io->note('Directory `' . $options->directory . '` looks like an existing Git repository.');
+        } catch (GitException $e) {
+            $io->note('Interpreting directory `' . $options->directory . '` as not a Git repository, cloning...');
+
+            $git->cloneRepository(
+                $mergeRequestCreateEvent->getGitUrl(),
+                $options->directory,
+                [
+                    '-b' => $mergeRequestCreateEvent->getGitBranch(),
+                ],
+            );
+
+            $io->success('Done');
+
+            return Command::SUCCESS;
+        }
+
+        // Check branch name.
+        $branch = $mergeRequestCreateEvent->getGitBranch();
+        while ($gitIo->branchExists($branch)) {
+            $branch = $io->ask(sprintf('Local branch with name `%s` already exists. Enter new branch name:', $branch), $branch);
+        }
+
+        $remotes = $gitIo->execute('remote');
+        $remoteUrls = [];
+        foreach ($remotes as $remoteName) {
+            $urls = array_fill_keys(
+                $gitIo->execute('remote', 'get-url', $remoteName),
+                $remoteName,
+            );
+            $remoteUrls = array_merge($remoteUrls, $urls);
+        }
+
+        if ($options->isHttp && isset($remoteUrls[$mergeRequestCreateEvent->getGitHttpUrl()])) {
+            $gitUrl = $mergeRequestCreateEvent->getGitHttpUrl();
+            $remoteName = $remoteUrls[$gitUrl];
+            $io->note(sprintf('Found existing HTTP remote for this merge request: %s @ %s', $remoteName, $gitUrl));
+        } elseif (!$options->isHttp && isset($remoteUrls[$mergeRequestCreateEvent->getGitUrl()])) {
+            $gitUrl = $mergeRequestCreateEvent->getGitUrl();
+            $remoteName = $remoteUrls[$gitUrl];
+            $io->note(sprintf('Found existing SSH remote for this merge request: %s @ %s', $remoteName, $gitUrl));
+        } else {
+            $gitUrl = $options->isHttp ? $mergeRequestCreateEvent->getGitHttpUrl() : $mergeRequestCreateEvent->getGitUrl();
+            $io->note('No existing HTTP remote for this merge request found.');
+
+            // Generate a unique remote name.
+            $remoteName = null;
+            $suffix = '';
+            while (null === $remoteName || in_array($remoteName, $remotes, true)) {
+                $remoteName = $mergeRequestCreateEvent->project() . '-' . $issue->id() . $suffix;
+                // Create a suffix for next run if the first attempt doesn't succeed:
+                $suffix = '-' . time();
+            }
+
+            $io->note(sprintf('Setting up new remote: %s @ %s', $remoteName, $gitUrl));
+            $gitIo->execute('remote', 'add', $remoteName, $gitUrl);
+        }
+
+        $io->note(sprintf('Fetching remote: %s @ %s', $remoteName, $gitUrl));
+        $gitIo->execute('fetch', $remoteName);
+
+        // Check if the SSH or HTTP URL for the MR we want to check out is already a remote:
+        $io->note(sprintf('Checking out branch: %s', $branch));
+        $gitIo->execute('checkout', '-b', "$branch", '--track', sprintf('%s/%s', $remoteName, $mergeRequestCreateEvent->getGitBranch()));
+
+        $io->success('Done');
+
+        return Command::SUCCESS;
+    }
+}
